% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOSCI220-package.R
\docType{package}
\name{BIOSCI220-package}
\alias{BIOSCI220}
\alias{BIOSCI220-package}
\title{BIOSCI220: Helper functions for BIOSCI 220 at the University of Auckland}
\description{
Helper functions for BIOSCI 220 at the University of Auckland
}
\author{
\strong{Maintainer}: Evelyn Jade \email{evelyn.jade@auckland.ac.nz}

Other contributors:
\itemize{
  \item Jenn Jury \email{jenn.jury@auckland.ac.nz} [copyright holder]
}

}
\keyword{internal}
