% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_11_lab.R
\name{calculate_active_recovered_cases}
\alias{calculate_active_recovered_cases}
\title{Calculate the number of active and recovered cases from daily new cases}
\usage{
calculate_active_recovered_cases(my_country_data)
}
\arguments{
\item{my_country_data}{`dataframe`. The covid data from Our World In Data, filtered to be from just one location/country.}
}
\value{
`dataframe`. A new dataframe of location, date, current active cases (new cases from the last 14 days), and current recovered cases (cases from before the last 14 days).
}
\description{
This function uses the daily new case numbers to calculate "active" cases (from the last 14 days)
and "recovered" cases (from before the last 14 days), in a dataframe that [ggplot2::ggplot()] can read.
}
