% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_11_lab.R
\name{ggplot_nz_example}
\alias{ggplot_nz_example}
\title{Make the example New Zealand plot}
\usage{
ggplot_nz_example(example_graph_data)
}
\arguments{
\item{example_graph_data}{`dataframe`. The active and recovered cases data for New Zealand, obtained from [calculate_active_recovered_cases()].}
}
\value{
`ggplot`. The example New Zealand plot we want you to replicate for your country of choice.
}
\description{
This function creates the example New Zealand active and recovered cases plot we want you to replicate.
}
